# build/tmp/kapt3/classes/main/META-INF/swagger/api-pix-1.0.yml
openapi: 3.0.1
info:
  title: api-pix
  version: "1.0"
paths:
  /client/{client-id}:
    get:
      operationId: findByClientId
      parameters:
      - name: client-id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: findByClientId 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'
  /pix/{client-id}:
    post:
      operationId: createPixRamdom
      parameters:
      - name: client-id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        "200":
          description: createPixRamdom 200 response
  /pix/{client-id}/cpf:
    post:
      operationId: createPixCpf
      parameters:
      - name: client-id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - request
              type: object
              properties:
                request:
                  $ref: '#/components/schemas/RegisterPixCpfFormRequest'
        required: true
      responses:
        "200":
          description: createPixCpf 200 response
  /pix/{client-id}/email:
    post:
      operationId: createPixEmail
      parameters:
      - name: client-id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPixEmailFormRequest'
        required: true
      responses:
        "200":
          description: createPixEmail 200 response
  /pix/{client-id}/phone:
    post:
      operationId: createPixPhone
      parameters:
      - name: client-id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPixPhoneFormRequest'
        required: true
      responses:
        "200":
          description: createPixPhone 200 response
components:
  schemas:
    Object:
      type: object
    RegisterPixCpfFormRequest:
      required:
      - key
      type: object
      properties:
        key:
          minLength: 1
          type: string
    RegisterPixEmailFormRequest:
      required:
      - emailKeyWord
      type: object
      properties:
        emailKeyWord:
          minLength: 1
          type: string
          format: email
    RegisterPixPhoneFormRequest:
      required:
      - phoneKeyWord
      type: object
      properties:
        phoneKeyWord:
          type: string